tweaks to make compiler happy
git-svn-id: svn+ssh://frodo/cluster/svn/projects/slash_nara@10544 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ include/fid.h
@@ -42,8 +42,8 @@ struct slash_fidgen;
 #define	SLASH_ID_SITE_BITS	10
 #define	SLASH_ID_FID_BITS	50
 
-#define SLFIDF_HIDE_DENTRY	(1 << 0)	/* keep but hide an entry until its log arrives */
-#define SLFIDF_LOCAL_DENTRY	(1 << 1)	/* don't expose to external nodes */
+#define SLFIDF_HIDE_DENTRY	(uint64_t)(1 << 0)	/* keep but hide an entry until its log arrives */
+#define SLFIDF_LOCAL_DENTRY	(uint64_t)(1 << 1)	/* don't expose to external nodes */
 
 typedef uint64_t slfid_t;
 typedef uint64_t slfgen_t;
